requirements.txt
fastapi
uvicorn
telethon
requests
python-dotenv
pillow
skimage
opencv-python-headless
numpy